Add ability to set background canvas color
