+++ Modules/MinoriEditorShell.VirtualCanvas/Platforms/Wpf/Controls/MesContentCanvas.cs
@@ -1,6 +1,7 @@
 ﻿using MinoriEditorShell.VirtualCanvas.Services;
 using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.Text;
 using System.Windows.Controls;
 
@@ -11,5 +12,16 @@ namespace MinoriEditorShell.VirtualCanvas.Platforms.Wpf.Controls
     /// </summary>
     public class MesContentCanvas : Canvas, IMesContentCanvas
     {
+        /// <summary>
+        /// Sets the background color of the canvas object
+        /// </summary>
+        /// <param name="color"></param>
+        public void SetCanvasBackgroundColor(Color color)
+        {
+            System.Windows.Media.Color newcolor = 
+                System.Windows.Media.Color.FromArgb(color.A, color.R, color.G, color.B);
+            System.Windows.Media.Brush brush = new System.Windows.Media.SolidColorBrush(newcolor);
+            Background = brush;
+        }
     }
 }
