+++ Modules/MinoriEditorShell.VirtualCanvas/Services/IMesContentCanvas.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.Text;
 
 namespace MinoriEditorShell.VirtualCanvas.Services
@@ -10,5 +11,10 @@ namespace MinoriEditorShell.VirtualCanvas.Services
     /// </summary>
     public interface IMesContentCanvas
     {
+        /// <summary>
+        /// Sets the background color of the canvas object
+        /// </summary>
+        /// <param name="color"></param>
+        void SetCanvasBackgroundColor(Color color);
     }
 }
