+++ Modules/MinoriEditorShell.VirtualCanvas/Services/IMesVirtualCanvasControl.cs
@@ -4,6 +4,17 @@ using System.Drawing;
 
 namespace MinoriEditorShell.VirtualCanvas.Services
 {
+    public class VisualChangeEventArgs : EventArgs
+    {
+        public Int32 Added { get; set; }
+        public Int32 Removed { get; set; }
+        public VisualChangeEventArgs(Int32 added, Int32 removed)
+        {
+            Added = added;
+            Removed = removed;
+        }
+    }
+
     public interface IMesVirtualCanvasControl
     {
         IMesContentCanvas ContentCanvas { get; }
@@ -14,6 +25,9 @@ namespace MinoriEditorShell.VirtualCanvas.Services
 
         ObservableCollection<IMesVirtualChild> VirtualChildren { get; }
         SizeF Extent { get; }
+        Int32 LiveVisualCount { get; }
+
+        event EventHandler<VisualChangeEventArgs> VisualsChanged;
 
         void AddVirtualChild(IMesVirtualChild shape);
     }
